% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{pkg_options}
\alias{pkg_options}
\title{Set or get options for this package}
\usage{
pkg_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for this package
}
\section{Supported options}{

The following options (listed with their defailt values) are supported:
\itemize{

\item{conventional.color="brown"}
\item{renewable.color="green"}
\item{storage_discharging.color="lightgreen"}
\item{storage_loading.color="skyblue"}
\item{curtailment.color="red"}
\item{demand.color="blue"}

}
}

